The Fix, the Bug and the Error

+++ src/main/java/plugin/database/wrappers/PlayerData.java
@@ -111,7 +111,7 @@ public class PlayerData {
         commit();
     }
     //mutators
-    public void increasePlaytime(int value) {
+    public void increasePlaytime(long value) {
         collection.playtime += value;
         commit();
     }
